globallib +ratings +sessions +adding to local lib

+++ Social_Library/Social_Library-war/src/java/com/sociallibrary/crud/LibraryCRUD.java
@@ -60,7 +60,7 @@ public class LibraryCRUD implements ILibraryCRUD {
         }
     }
 
-    public Library readLibrary(int id)
+    public Library readLibrary(long id)
     {
         BasicConfigurator.configure();
          ResultSet rs=null;
